Port to python3
Tested with python3.8.6. All functional tests work, but I was not able
to run the performance tests. They also relied on time.clock() in several
places. This function is deprecated since version 3.3 and was removed
in version 3.8, because of highly platform specific behavior.
I replaced it with time.process_time() as there is no direct replacement.
This invalidates all measurement values in the performance tests.